requirements_full_project.txt
flask
flask-cors
google-generativeai==0.5.2
Pillow
torch
opencv-python
numpy
python-dotenv
requests
